Mania Bot/requirements.txt
astroid==2.4.2
colorama==0.4.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
MouseInfo==0.1.3
mss==2.0.22
numpy==1.19.1
opencv-python==4.4.0.40
Pillow==7.2.0
PyAutoGUI==0.9.50
PyGetWindow==0.0.8
pylint==2.5.3
PyMsgBox==1.0.8
pyperclip==1.8.0
PyRect==0.1.4
PyScreeze==0.1.26
PyTweening==1.0.3
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
